Use TypeScript and adapt config and tests
