+++ package.json
@@ -6,11 +6,11 @@
   },
   "version": "1.0.0",
   "description": "A simple MJML component to make static bar charts",
+  "source": "src/index.ts",
   "main": "lib/index.js",
   "scripts": {
-    "start": "gulp watch",
-    "preprepare": "eslint components",
-    "prepare": "gulp build",
+    "watch": "tsc --watch",
+    "build": "tsc",
     "test": "jest tests",
     "release": "scripts/release.sh"
   },
@@ -18,23 +18,19 @@
   "homepage": "https://github.com/Freezystem/mjml-bar-chart",
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.10.4",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/register": "^7.10.4",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.4.0",
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "fancy-log": "^1.3.3",
-    "gulp": "^4.0.2",
-    "gulp-babel": "^8.0.0",
-    "gulp-watch": "^5.0.1",
-    "jest": "^27.0.6"
+    "@types/jest": "^27.0.3",
+    "@types/mjml": "^4.7.0",
+    "@types/mjml-core": "^4.7.1",
+    "jest": "^27.4.5",
+    "ts-jest": "^27.1.2",
+    "typescript": "^4.5.4"
   },
   "dependencies": {
-    "mjml": "^4.10.2",
-    "mjml-core": "^4.10.2"
+    "mjml": "^4.11.0",
+    "mjml-core": "^4.11.0"
+  },
+  "jest": {
+    "preset": "ts-jest/presets/js-with-ts-esm",
+    "testEnvironment": "node"
   }
 }
